Prototype of writing log entries into CloudWatch.
Not finished yet: need to make the log group name configurable and make
it create a log stream named after the EC2 instance id.

Also need to make the state management more robust so that we're more
likely to successfully write out the state if we get a TERM or INT signal.