Terminate gracefully on SIGTERM or SIGINT
We need to make sure we write the latest resume token when we exit so
that we can resume next time we run. To make this more likely we will
respond gracefully to SIGTERM and SIGINT so that we'll deplete the
buffers, flush everything we got to the Cloudwatch Logs API and then
write the final state to the state file before we exit.

Previously we might not have done this correctly if we happened to get
a signal during a write to cloudwatch, since cloudwatch would return to us
a new token but we'd never get a chance to write it to the state file
before we exit.