Add support for priority filtering
In our environment we sometimes run things with DEBUG mode.
In these cases journald is filled with a bunch of messages, irrelevant
for the normal work mode, but reading them and pushing to CloudWatch
uses considerable amount of io/cpu on the host system.

* Add a configuration parameter "log_priority" to filter log messages
similar to the "journalctl -p <prio>" command.

Signed-off-by: Alex Lourie <djay.il@gmail.com>
Signed-off-by: Alex Lourie <alex@instaclustr.com>