create log group if resource does not exist
This adds an extra step to the startup and pushEvents that attempts to
create the log group in cloudwatch, similar to how the `createStream`
function works. However, instead of attempting to push the logs write
after, fall through to create the stream as well (because obviously it
won't exist yet either).